requirements.txt
fastapi
uvicorn
pydantic
libvirt-python
python-dotenv
prometheus-client
psutil
asyncssh
